last

+++ db/db_data/likes.py
@@ -11,4 +11,4 @@ class Likes(SqlAlchemyBase):
     users_mutual = sqlalchemy.Column(sqlalchemy.String, default='')
     sbd_likes_user = sqlalchemy.Column(sqlalchemy.String, default='')
     sbd_unlikes_user = sqlalchemy.Column(sqlalchemy.String, default='')
-    rating = sqlalchemy.Column(sqlalchemy.Float, default=0.0)
+    rating = sqlalchemy.Column(sqlalchemy.Integer, default=0)
